+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "next-starter",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "description": "A minimal Next.js starter template preconfigured with TypeScript, ESLint, Prettier, and Husky.",
   "author": "Christopher Carpenter <mrchriscarpenter@outlook.com>",
   "license": "MIT",
@@ -19,19 +19,19 @@
     "**/*": "prettier --write --ignore-unknown"
   },
   "dependencies": {
-    "next": "13.2.4",
+    "next": "13.4.19",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "@types/node": "^18.15.8",
-    "@types/react": "^18.0.29",
-    "@types/react-dom": "^18.0.11",
+    "@types/node": "^20.5.6",
+    "@types/react": "^18.2.21",
+    "@types/react-dom": "^18.2.7",
     "@typescript-eslint/eslint-plugin": "^5.56.0",
     "@typescript-eslint/parser": "^5.32.0",
-    "eslint": "8.36.0",
+    "eslint": "8.47.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-next": "13.2.4",
+    "eslint-config-next": "13.4.19",
     "eslint-config-prettier": "^8.8.0",
     "eslint-import-resolver-typescript": "^3.4.0",
     "eslint-plugin-import": "^2.27.5",
@@ -41,6 +41,6 @@
     "husky": "^8.0.3",
     "lint-staged": "^13.2.0",
     "prettier": "2.8.7",
-    "typescript": "5.0.2"
+    "typescript": "5.2.2"
   }
 }
